Add closePopover
